Remove v3 endpoint for Beckman
