+++ lighthouse/__init__.py
@@ -7,10 +7,6 @@ from flask_apscheduler import APScheduler
 
 from lighthouse.hooks.cherrypick_test_data import inserted_cherrypick_test_data_hook
 from lighthouse.hooks.events import insert_events_hook, inserted_events_hook
-from lighthouse.routes import eve_routes
-from lighthouse.routes.v1 import beckman_routes as v1_beckman_routes
-from lighthouse.routes.v1 import routes as v1_routes
-from lighthouse.routes.v3 import beckman_routes as v3_beckman_routes
 from lighthouse.validator import LighthouseValidator
 
 scheduler = APScheduler()
@@ -44,15 +40,19 @@ def create_app() -> Eve:
 
 
 def setup_routes(app):
+    from lighthouse.routes import eve_routes
+    from lighthouse.routes.v1 import beckman_routes as v1_beckman_routes
+    from lighthouse.routes.v1 import routes as v1_routes
+
     # When registering blueprints, do so both in the root and in versioned endpoints such as /v1.
     # Register latest version routes at the root of the API -- note that Eve automatically registered at the root.
     app.register_blueprint(v1_routes.bp, name="root_routes")
-    app.register_blueprint(v3_beckman_routes.bp, name="root_beckman_routes")
+    app.register_blueprint(v1_beckman_routes.bp, name="root_beckman_routes")
 
     # Register /v1 routes
     app.register_blueprint(eve_routes.bp, url_prefix="/v1")
     app.register_blueprint(v1_routes.bp, url_prefix="/v1")
     app.register_blueprint(v1_beckman_routes.bp, url_prefix="/v1")
 
-    # Register /v3 routes (there is no /v2)
-    app.register_blueprint(v3_beckman_routes.bp, url_prefix="/v3")
+    # Register further version of routes
+    # For example: app.register_blueprint(v2_routes.bp, url_prefix="/v2")
